Allow perf_data_converter to read file size larger than 4GB
This changes the raw_size to uint64_t to allow quipper to take larger files. Note the underlying protobuf size is still capped at 2GB, which means this will not take arbitrary large files.

PiperOrigin-RevId: 458324805

+++ src/perf_data_converter.h
@@ -130,7 +130,7 @@ using ProcessProfiles = std::vector<std::unique_ptr<ProcessProfile>>;
 //
 // Returns a vector of process profiles, empty if any error occurs.
 extern ProcessProfiles RawPerfDataToProfiles(
-    const void* raw, int raw_size,
+    const void* raw, uint64_t raw_size,
     const std::map<std::string, std::string>& build_ids,
     uint32 sample_labels = kNoLabels, uint32 options = kGroupByPids,
     const std::map<uint32, std::string>& thread_types = {});
